Rename lib/util to lib/utils
