+++ lib/api/controllers/records.js
@@ -1,9 +1,7 @@
-'use strict'
-
 const mongoose = require('mongoose')
 const {clone} = require('lodash')
 const search = require('../../search')
-const {proxyThumbnail} = require('../../util/thumbnail-proxy')
+const {proxyThumbnail} = require('../../utils/thumbnail-proxy')
 
 const ConsolidatedRecord = mongoose.model('ConsolidatedRecord')
 const RecordRevision = mongoose.model('RecordRevision')
