+++ lib/api/routes/links.js
@@ -6,7 +6,7 @@ const got = require('got')
 const pump = require('pump')
 
 const {getLink} = require('../../services/link-proxy')
-const {createTempDirectory} = require('../../util/tempfile')
+const {createTempDirectory} = require('../../utils/tempfile')
 const ogr2ogr = require('../../extract/ogr2ogr')
 
 const am = require('../middlewares/async')
