+++ lib/metadata/common/links.js
@@ -1,8 +1,10 @@
 'use strict'
 
-const {cleanURL, isPublicURL} = require('../../util/url')
 const URI = require('urijs')
 const {forEach} = require('lodash')
+
+const {cleanURL, isPublicURL} = require('../../utils/url')
+
 const {getNormalizedWfsServiceLocation} = require('./services')
 
 class OnlineResource {
