+++ lib/metadata/iso/services.js
@@ -2,7 +2,7 @@
 
 const {get, pick, chain} = require('lodash')
 
-const {isPublicURL} = require('../../util/url')
+const {isPublicURL} = require('../../utils/url')
 const {getNormalizedWfsServiceLocation} = require('../common/services')
 
 const {getAllKeywords} = require('./keywords')
