+++ lib/metadata/iso/thumbnails.js
@@ -1,7 +1,6 @@
-'use strict'
-
-const {isPublicURL} = require('../../util/url')
 const hasha = require('hasha')
+
+const {isPublicURL} = require('../../utils/url')
 
 exports.prepareThumbnails = function (graphicOverviews = []) {
   return graphicOverviews
