+++ lib/utils/thumbnail-proxy.js
